+++ Source/EventFlow.Tests/Exploration/RegisterSubscribersExplorationTests.cs
@@ -26,6 +26,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using EventFlow.Aggregates;
+using EventFlow.Configuration;
 using EventFlow.Core;
 using EventFlow.Extensions;
 using EventFlow.Subscribers;
@@ -33,6 +34,7 @@ using EventFlow.TestHelpers;
 using EventFlow.TestHelpers.Aggregates;
 using EventFlow.TestHelpers.Aggregates.Commands;
 using EventFlow.TestHelpers.Aggregates.Events;
+using EventFlow.TestHelpers.Aggregates.Queries;
 using EventFlow.TestHelpers.Aggregates.ValueObjects;
 using FluentAssertions;
 using NUnit.Framework;
@@ -51,6 +53,7 @@ namespace EventFlow.Tests.Exploration
             using (var resolver =  register(EventFlowOptions.New)
                 .AddCommands(typeof(ThingyPingCommand))
                 .AddCommandHandlers(typeof(ThingyPingCommandHandler))
+                .RegisterServices(sr => sr.Register<IScopedContext, ScopedContext>(Lifetime.Scoped))
                 .AddEvents(typeof(ThingyPingEvent))
                 .Configure(c => c.IsAsynchronousSubscribersEnabled = true)
                 .CreateResolver(false))