+++ Source/EventFlow/Core/IoC/EventFlowIoCResolver.cs
@@ -34,16 +34,21 @@ namespace EventFlow.Core.IoC
     internal class EventFlowIoCResolver : IRootResolver
     {
         private readonly IReadOnlyDictionary<Type, List<Registration>> _registrations;
-        private readonly bool _dispose;
+        private readonly bool _disposeSingletons;
         private readonly IResolverContext _resolverContext;
         private readonly Type[] _emptyTypeArray = {};
+        private readonly List<IDisposable> _uniqueDisposables = new List<IDisposable>();
+        private readonly ConcurrentDictionary<int, object> _singletons;
+        private readonly ConcurrentDictionary<int, object> _scoped = new ConcurrentDictionary<int, object>();
 
         public EventFlowIoCResolver(
+            ConcurrentDictionary<int, object> singletons,
             IReadOnlyDictionary<Type, List<Registration>> registrations,
-            bool dispose)
+            bool disposeSingletons)
         {
+            _singletons = singletons;
             _registrations = registrations;
-            _dispose = dispose;
+            _disposeSingletons = disposeSingletons;
 
             _resolverContext = new ResolverContext(this);
         }
@@ -55,7 +60,6 @@ namespace EventFlow.Core.IoC
 
         public object Resolve(Type serviceType)
         {
-            List<Registration> registrations;
             var genericArguments = _emptyTypeArray;
             var typeInfo = serviceType.GetTypeInfo();
 
@@ -68,7 +72,9 @@ namespace EventFlow.Core.IoC
                 typeInfo = serviceType.GetTypeInfo();
             }
 
-            if (!_registrations.TryGetValue(serviceType, out registrations))
+            IRegistration registration;
+
+            if (!_registrations.TryGetValue(serviceType, out var registrations))
             {
                 if (serviceType == typeof(IResolver))
                 {
@@ -77,29 +83,26 @@ namespace EventFlow.Core.IoC
 
                 if (serviceType == typeof(IScopeResolver))
                 {
-                    return BeginScope();
+                    return new EventFlowIoCResolver(_singletons, _registrations, false);
                 }
 
                 if (typeInfo.IsGenericType && typeInfo.GetGenericTypeDefinition() == typeof(IEnumerable<>))
                 {
-                    return EnumerableResolvers
-                        .GetOrAdd(serviceType, t => new EnumerableResolver(t))
-                        .Create(_resolverContext);
+                    registration = EnumerableResolvers.GetOrAdd(serviceType, t => new EnumerableResolver(t));
+                    return Create(registration);
                 }
 
                 throw new Exception($"Type {serviceType.PrettyPrint()} is not registered");
             }
 
-            return registrations.First().Create(_resolverContext, genericArguments);
+            registration = registrations.First();
+            return Create(registration, genericArguments);
         }
-
 
         public IEnumerable<object> ResolveAll(Type serviceType)
         {
-            List<Registration> registrations;
-
-            return _registrations.TryGetValue(serviceType, out registrations)
-                ? registrations.Select(r => r.Create(_resolverContext, _emptyTypeArray))
+            return _registrations.TryGetValue(serviceType, out var registrations)
+                ? registrations.Select(r => Create(r, _emptyTypeArray))
                 : Enumerable.Empty<object>();
         }
 
@@ -117,24 +120,76 @@ namespace EventFlow.Core.IoC
 
         public void Dispose()
         {
-            if (!_dispose) return;
-
-            foreach (var registration in _registrations.Values.SelectMany(r => r))
+            foreach (var disposable in _uniqueDisposables)
             {
-                registration.Dispose();
+                disposable.Dispose();
             }
+            _uniqueDisposables.Clear();
+
+            foreach (var scoped in _scoped.Values)
+            {
+                (scoped as IDisposable)?.Dispose();
+            }
+            _scoped.Clear();
+
+            if (!_disposeSingletons)
+            {
+                return;
+            }
+
+            foreach (var singleton in _singletons.Values)
+            {
+                (singleton as IDisposable)?.Dispose();
+            }
+            _singletons.Clear();
         }
 
         public IScopeResolver BeginScope()
         {
-            return new EventFlowIoCResolver(_registrations, false);
+            return new EventFlowIoCResolver(_singletons, _registrations, false);
         }
 
-        private static readonly ConcurrentDictionary<Type, EnumerableResolver> EnumerableResolvers = new ConcurrentDictionary<Type, EnumerableResolver>();
-
-        private class EnumerableResolver
+        private object Create(IRegistration registration, params Type[] genericTypeArguments)
         {
-            private readonly Type _genericArgument;
+            switch (registration.Lifetime)
+            {
+                case Lifetime.AlwaysUnique:
+                    var service = registration.Create(_resolverContext, genericTypeArguments);
+                    if (service is IDisposable disposable)
+                    {
+                        _uniqueDisposables.Add(disposable);
+                    }
+                    return service;
+
+                case Lifetime.Singleton:
+                    return _singletons.GetOrAdd(
+                        CombineHash(registration.ServiceType, genericTypeArguments),
+                        _ => registration.Create(_resolverContext, genericTypeArguments));
+
+                case Lifetime.Scoped:
+                    return _scoped.GetOrAdd(
+                        CombineHash(registration.ServiceType, genericTypeArguments),
+                        _ => registration.Create(_resolverContext, genericTypeArguments));
+
+                default:
+                    throw new ArgumentOutOfRangeException();
+            }
+        }
+
+        private static int CombineHash(Type serviceType, params Type[] genericTypeArguments)
+        {
+            return genericTypeArguments.Aggregate(
+                serviceType.GetHashCode(),
+                (i, t) => HashHelper.Combine(i, t.GetHashCode()));
+        }
+
+        private static readonly ConcurrentDictionary<Type, IRegistration> EnumerableResolvers = new ConcurrentDictionary<Type, IRegistration>();
+
+        private class EnumerableResolver : IRegistration
+        {
+            public Type ServiceType { get; }
+            public Lifetime Lifetime => Lifetime.AlwaysUnique;
+
             private readonly MethodInfo _methodInfo;
 
             public EnumerableResolver(
@@ -145,8 +200,8 @@ namespace EventFlow.Core.IoC
                     .GetMethods(BindingFlags.Static | BindingFlags.NonPublic)
                     .Single(mi => mi.Name == nameof(ResolveEnumerable));
 
-                _genericArgument = typeInfo.GenericTypeArguments.Single();
-                _methodInfo = genericMethodInfo.MakeGenericMethod(_genericArgument);
+                ServiceType = typeInfo.GenericTypeArguments.Single();
+                _methodInfo = genericMethodInfo.MakeGenericMethod(ServiceType);
             }
 
             private static IEnumerable<T> ResolveEnumerable<T>(IResolverContext resolverContext, Type serviceType)
@@ -156,9 +211,9 @@ namespace EventFlow.Core.IoC
                     .Select(s => (T)s);
             }
 
-            public object Create(IResolverContext resolverContext)
+            public object Create(IResolverContext resolverContext, Type[] genericTypeArguments)
             {
-                return _methodInfo.Invoke(null, new object[] { resolverContext, _genericArgument });
+                return _methodInfo.Invoke(null, new object[] { resolverContext, ServiceType });
             }
         }
     }